modules/locationConsumer/requirements.txt
psycopg2-binary==2.8.5
SQLAlchemy==1.3.19
kafka-python